backend/requirements.txt
flask
flask-cors
openai
pdfminer.six
python-docx
scikit-learn
